requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
imblearn
pymongo
python-dotenv
PyYAML
certifi

-e . # editable mode